% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addmath.R
\name{addmath}
\alias{addmath}
\title{add two numeric vectors of same length}
\usage{
addmath(a, b)
}
\arguments{
\item{a}{A numeric vector of length n}

\item{b}{A numeric vector of length n}
}
\value{
Returns a numeric vector of length n
}
\description{
add two numeric vectors of same length
}
\examples{
addmath(c(1,2),c(3,4))
addmath(1,2)
}
